Adds tests for literals in a new module
