+++ tests/dataframe_cookbook/test_python_list_cols.py
@@ -58,24 +58,6 @@ def parametrize_udfs(name: str, funcs: Callable | list[Callable], return_type: t
 ###
 
 
-def test_pyobj_literal_column(daft_df, service_requests_csv_pd_df):
-    """Creating a new column that is derived from (1 + other_column) and retrieving the top N results"""
-    daft_df = daft_df.with_column("literal_col", lit({"foo": "bar"}))
-    daft_pd_df = daft_df.to_pandas()
-    service_requests_csv_pd_df["literal_col"] = pd.Series(
-        {"foo": "bar"} for _ in range(len(service_requests_csv_pd_df))
-    )
-    assert_df_equals(daft_pd_df, service_requests_csv_pd_df)
-
-
-def test_literal_column_computation(daft_df, service_requests_csv_pd_df):
-    """Creating a new column that is derived from (1 + other_column) and retrieving the top N results"""
-    daft_df = daft_df.with_column("literal_col", lit({"foo": "bar"}).apply(lambda d: d["foo"], return_type=str))
-    daft_pd_df = daft_df.to_pandas()
-    service_requests_csv_pd_df["literal_col"] = "bar"
-    assert_df_equals(daft_pd_df, service_requests_csv_pd_df)
-
-
 @pytest.mark.parametrize("repartition_nparts", [1, 5, 6, 10, 11])
 def test_python_dict(repartition_nparts):
     data = {"id": [i for i in range(10)], "dicts": [{"foo": i} for i in range(10)]}
